+++ eth_defi/uniswap_v2/pair.py
@@ -52,6 +52,11 @@ class PairDetails:
         """Get pair contract address"""
         return self.contract.address
 
+    @property
+    def checksum_free_address(self) -> str:
+        """Get pair contract address, all lowercase."""
+        return self.contract.address.lower()
+
     def get_base_token(self):
         """Get human-ordered base token."""
         assert self.reverse_token_order is not None, "Reverse token order flag must be check before this operation is possible"